Projet_z/requirements.txt
crewai>=0.30.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0
requests>=2.31.0
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0